// components/select-user/interface.ts
export type NodeType = 'USER' | 'ORG' | 'DEPT' | 'GROUP_DEPT' | 'GROUP' | 'TAG';
// ...
export interface ItreeItem {
  id: string;
  key: string;
  name: string;
  label: any;
  nodeType: string;
  orgId?: string;
  type?: NodeType;
  deptType?:
    | 0 // 基础校区
    | 1; // 自定义校区
  // 所属部门名称（用于区分不同部门下的同一人，对应 strictUser 配置）
  deptName?: string;
  // 下属的人员数量，用于右侧统计
  userCount?: number;
  // 直接子节点
  children?: ItreeItem[];
  // 是否可选中
  checkable?: boolean;
  // 是否叶子节点
  isLeaf?: boolean;
  // 图标
  icon?: any;
  contactType?: string;
  count?: number;
}